Refactor Topic tests to abstract common tests and highlight differences
