+++ hazelcast-hibernate5/src/test/java/com/hazelcast/hibernate/CacheHitMissNonStrictTest.java
@@ -39,11 +39,11 @@ import static org.junit.Assert.assertEquals;
  */
 @RunWith(HazelcastSerialClassRunner.class)
 @Category({QuickTest.class, ParallelTest.class})
-public class CacheHitMissNonStrictTest
-        extends HibernateStatisticsTestSupport {
+public class CacheHitMissNonStrictTest extends HibernateStatisticsTestSupport {
 
-    protected String getCacheStrategy() {
-        return AccessType.NONSTRICT_READ_WRITE.getExternalName();
+    @Override
+    protected AccessType getCacheStrategy() {
+        return AccessType.NONSTRICT_READ_WRITE;
     }
 
     @Override
