+++ hazelcast-hibernate5/src/test/java/com/hazelcast/hibernate/TopicReadWriteTest.java
@@ -16,89 +16,18 @@
 
 package com.hazelcast.hibernate;
 
-import com.hazelcast.hibernate.entity.AnnotatedEntity;
-import com.hazelcast.hibernate.entity.DummyEntity;
-import com.hazelcast.hibernate.entity.DummyProperty;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.SlowTest;
+import org.hibernate.Query;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
-import org.hibernate.cache.spi.access.AccessType;
-import org.hibernate.cfg.Environment;
-
-import org.hibernate.Query;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import java.util.*;
-
 @RunWith(HazelcastSerialClassRunner.class)
 @Category(SlowTest.class)
-public class TopicReadWriteTest extends HibernateTopicTestSupport {
-
-    public AccessType defaultAccessType = AccessType.READ_WRITE;
-
-    @Test
-    public void testReplaceCollection() {
-        insertDummyEntities(sf, 2, 4);
-
-        Session session = sf.openSession();
-        Transaction transaction = session.beginTransaction();
-
-        DummyProperty property = new DummyProperty("somekey");
-        session.save(property);
-
-        DummyEntity entity = session.get(DummyEntity.class, 1L);
-        HashSet<DummyProperty> updatedProperties = new HashSet<DummyProperty>();
-        updatedProperties.add(property);
-        entity.setProperties(updatedProperties);
-
-        session.update(entity);
-
-        transaction.commit();
-        session.close();
-
-        assertTopicNotifications(1, CACHE_ENTITY);
-        assertTopicNotifications(2, CACHE_ENTITY_PROPERTIES);
-        assertTopicNotifications(18, CACHE_TIMESTAMPS_REGION);
-    }
-
-    @Test
-    public void testUpdateOneEntityByNaturalId() {
-        insertAnnotatedEntities(sf, 2);
-
-        Session session = sf.openSession();
-        Transaction tx = session.beginTransaction();
-
-        AnnotatedEntity toBeUpdated = session.byNaturalId(AnnotatedEntity.class).using("title", "dummy:1").getReference();
-        toBeUpdated.setTitle("dummy101");
-        tx.commit();
-        session.close();
-
-        assertTopicNotifications(2, CACHE_ANNOTATED_ENTITY + "##NaturalId");
-        assertTopicNotifications(4, CACHE_TIMESTAMPS_REGION);
-    }
-
-    @Test
-    public void testUpdateEntitiesByNaturalId() {
-        insertAnnotatedEntities(sf, 2);
-
-        Session session = sf.openSession();
-        Transaction tx = session.beginTransaction();
-
-        AnnotatedEntity toBeUpdated = session.byNaturalId(AnnotatedEntity.class).using("title", "dummy:1").getReference();
-        toBeUpdated.setTitle("dummy101");
-
-        AnnotatedEntity toBeUpdated2 = session.byNaturalId(AnnotatedEntity.class).using("title", "dummy:0").getReference();
-        toBeUpdated2.setTitle("dummy100");
-        tx.commit();
-        session.close();
-
-        // 5 notifications = 1 eviction, plus for each update: one unlockItem and one afterUpdate
-        assertTopicNotifications(5, CACHE_ANNOTATED_ENTITY + "##NaturalId");
-        assertTopicNotifications(4, CACHE_TIMESTAMPS_REGION);
-    }
+public class TopicReadWriteTest extends TopicReadWriteTestSupport {
 
     @Test
     public void testUpdateQueryByNaturalId() {
@@ -111,93 +40,6 @@ public class TopicReadWriteTest extends HibernateTopicTestSupport {
     }
 
     @Test
-    public void testDeleteOneEntityByNaturalId() {
-        insertAnnotatedEntities(sf, 2);
-
-        Session session = sf.openSession();
-        Transaction tx = session.beginTransaction();
-
-        AnnotatedEntity toBeDeleted = session.byNaturalId(AnnotatedEntity.class).using("title", "dummy:1").getReference();
-        session.delete(toBeDeleted);
-        tx.commit();
-        session.close();
-
-        assertTopicNotifications(2, CACHE_ANNOTATED_ENTITY + "##NaturalId");
-        assertTopicNotifications(4, CACHE_TIMESTAMPS_REGION);
-    }
-
-    @Test
-    public void testDeleteEntitiesByNaturalId() {
-        insertAnnotatedEntities(sf, 2);
-
-        Session session = sf.openSession();
-        Transaction tx = session.beginTransaction();
-
-        AnnotatedEntity toBeDeleted = session.byNaturalId(AnnotatedEntity.class).using("title", "dummy:1").getReference();
-        session.delete(toBeDeleted);
-
-        AnnotatedEntity toBeDeleted2 = session.byNaturalId(AnnotatedEntity.class).using("title", "dummy:0").getReference();
-        session.delete(toBeDeleted2);
-        tx.commit();
-        session.close();
-
-        assertTopicNotifications(4, CACHE_ANNOTATED_ENTITY + "##NaturalId");
-        assertTopicNotifications(4, CACHE_TIMESTAMPS_REGION);
-    }
-
-    @Test
-    public void testDeleteOneEntity() throws Exception {
-        insertDummyEntities(sf, 1, 1);
-
-        deleteDummyEntity(sf, 0);
-
-        assertTopicNotifications(1, CACHE_ENTITY);
-        assertTopicNotifications(1, CACHE_ENTITY_PROPERTIES);
-        assertTopicNotifications(1, CACHE_PROPERTY);
-        assertTopicNotifications(9, CACHE_TIMESTAMPS_REGION);
-    }
-
-    @Test
-    public void testDeleteEntities() throws Exception {
-        insertDummyEntities(sf, 10, 4);
-
-        for (int i = 0; i < 3; i++) {
-            deleteDummyEntity(sf, i);
-        }
-
-        assertTopicNotifications(3, CACHE_ENTITY);
-        assertTopicNotifications(3, CACHE_ENTITY_PROPERTIES);
-        assertTopicNotifications(12, CACHE_PROPERTY);
-        assertTopicNotifications(67, CACHE_TIMESTAMPS_REGION);
-    }
-
-    @Test
-    public void testUpdateOneEntity() {
-        insertDummyEntities(sf, 10, 4);
-
-        getDummyEntities(sf, 10);
-
-        updateDummyEntityName(sf, 2, "updated");
-
-        assertTopicNotifications(1, CACHE_ENTITY);
-        assertTopicNotifications(0, CACHE_ENTITY_PROPERTIES);
-        assertTopicNotifications(0, CACHE_PROPERTY);
-        assertTopicNotifications(54, CACHE_TIMESTAMPS_REGION);
-    }
-
-    @Test
-    public void testUpdateEntities() {
-        insertDummyEntities(sf, 1, 10);
-
-        executeUpdateQuery(sf, "update DummyEntity set name = 'updated-name' where id < 2");
-
-        assertTopicNotifications(1, CACHE_ENTITY);
-        assertTopicNotifications(0, CACHE_ENTITY_PROPERTIES);
-        assertTopicNotifications(0, CACHE_PROPERTY);
-        assertTopicNotifications(15, CACHE_TIMESTAMPS_REGION);
-    }
-
-    @Test
     public void testUpdateEntitiesAndProperties() {
         insertDummyEntities(sf, 1, 10);
 
@@ -206,11 +48,11 @@ public class TopicReadWriteTest extends HibernateTopicTestSupport {
         try {
             session = sf.openSession();
             txn = session.beginTransaction();
-            Query query = session.createQuery( "update DummyEntity set name = 'updated-name' where id < 2");
+            Query query = session.createQuery("update DummyEntity set name = 'updated-name' where id < 2");
             query.setCacheable(true);
             query.executeUpdate();
 
-            Query query2 = session.createQuery( "update DummyProperty set version = version + 1");
+            Query query2 = session.createQuery("update DummyProperty set version = version + 1");
             query2.setCacheable(true);
             query2.executeUpdate();
 
@@ -238,11 +80,11 @@ public class TopicReadWriteTest extends HibernateTopicTestSupport {
         try {
             session = sf.openSession();
             txn = session.beginTransaction();
-            Query query = session.createQuery( "update DummyEntity set name = 'updated-name' where id = 0");
+            Query query = session.createQuery("update DummyEntity set name = 'updated-name' where id = 0");
             query.setCacheable(true);
             query.executeUpdate();
 
-            Query query2 = session.createQuery( "update DummyProperty set version = version + 1");
+            Query query2 = session.createQuery("update DummyProperty set version = version + 1");
             query2.setCacheable(true);
             query2.executeUpdate();
 
@@ -259,17 +101,5 @@ public class TopicReadWriteTest extends HibernateTopicTestSupport {
         assertTopicNotifications(1, CACHE_ENTITY_PROPERTIES);
         assertTopicNotifications(1, CACHE_PROPERTY);
         assertTopicNotifications(17, CACHE_TIMESTAMPS_REGION);
-    }
-
-    protected String getCacheStrategy() {
-        return defaultAccessType.getExternalName();
-    }
-
-    @Override
-    protected Properties getCacheProperties() {
-        Properties props = new Properties();
-        props.put("TestAccessType", defaultAccessType);
-        props.setProperty(Environment.CACHE_REGION_FACTORY, HazelcastLocalCacheRegionFactory.class.getName());
-        return props;
     }
 }
