+++ hazelcast-hibernate5/src/test/java/com/hazelcast/hibernate/TopicReadWriteTestSupport.java
@@ -1,4 +1,4 @@
-package com.hazelcast.hibernate;/*
+/*
  * Copyright (c) 2008-2016, Hazelcast, Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,31 +14,22 @@ package com.hazelcast.hibernate;/*
  * limitations under the License.
  */
 
+package com.hazelcast.hibernate;
+
 import com.hazelcast.hibernate.entity.AnnotatedEntity;
 import com.hazelcast.hibernate.entity.DummyEntity;
 import com.hazelcast.hibernate.entity.DummyProperty;
-import com.hazelcast.test.HazelcastSerialClassRunner;
-import com.hazelcast.test.annotation.SlowTest;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 import org.hibernate.cache.spi.access.AccessType;
-import org.hibernate.cfg.Environment;
-import org.hibernate.query.Query;
-import org.junit.Ignore;
+
 import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
 
-import java.util.HashSet;
-import java.util.Properties;
+import java.util.*;
 
-@RunWith(HazelcastSerialClassRunner.class)
-@Category(SlowTest.class)
-public class TopicReadOnlyTest extends HibernateTopicTestSupport {
+public abstract class TopicReadWriteTestSupport extends HibernateTopicTestSupport {
 
-    public AccessType defaultAccessType = AccessType.READ_ONLY;
-
-    @Test(expected = UnsupportedOperationException.class)
+    @Test
     public void testReplaceCollection() {
         insertDummyEntities(sf, 2, 4);
 
@@ -49,7 +40,7 @@ public class TopicReadOnlyTest extends HibernateTopicTestSupport {
         session.save(property);
 
         DummyEntity entity = session.get(DummyEntity.class, 1L);
-        HashSet<DummyProperty> updatedProperties = new HashSet<>();
+        HashSet<DummyProperty> updatedProperties = new HashSet<DummyProperty>();
         updatedProperties.add(property);
         entity.setProperties(updatedProperties);
 
@@ -57,6 +48,10 @@ public class TopicReadOnlyTest extends HibernateTopicTestSupport {
 
         transaction.commit();
         session.close();
+
+        assertTopicNotifications(1, CACHE_ENTITY);
+        assertTopicNotifications(2, CACHE_ENTITY_PROPERTIES);
+        assertTopicNotifications(18, CACHE_TIMESTAMPS_REGION);
     }
 
     @Test
@@ -92,19 +87,6 @@ public class TopicReadOnlyTest extends HibernateTopicTestSupport {
 
         // 5 notifications = 1 eviction, plus for each update: one unlockItem and one afterUpdate
         assertTopicNotifications(5, CACHE_ANNOTATED_ENTITY + "##NaturalId");
-        assertTopicNotifications(4, CACHE_TIMESTAMPS_REGION);
-    }
-
-    @Test
-    public void testUpdateQueryByNaturalId() {
-        insertAnnotatedEntities(sf, 2);
-
-        executeUpdateQuery(sf, "update AnnotatedEntity set title = 'updated-name' where title = 'dummy:1'");
-
-        // There are *2* topic notifications (compared to *1* on previous Hibernate versions):
-        // - removeAll is called after executing the update
-        // - unlockRegion is called after the transaction completes
-        assertTopicNotifications(2, CACHE_ANNOTATED_ENTITY + "##NaturalId");
         assertTopicNotifications(4, CACHE_TIMESTAMPS_REGION);
     }
 
@@ -149,11 +131,9 @@ public class TopicReadOnlyTest extends HibernateTopicTestSupport {
 
         deleteDummyEntity(sf, 0);
 
-        assertTopicNotifications(2, CACHE_ENTITY);
-        // There is only *1* topic notification (compared to *2* - which are both an evict by the key - on previous
-        // Hibernate versions)
+        assertTopicNotifications(1, CACHE_ENTITY);
         assertTopicNotifications(1, CACHE_ENTITY_PROPERTIES);
-        assertTopicNotifications(2, CACHE_PROPERTY);
+        assertTopicNotifications(1, CACHE_PROPERTY);
         assertTopicNotifications(9, CACHE_TIMESTAMPS_REGION);
     }
 
@@ -165,100 +145,40 @@ public class TopicReadOnlyTest extends HibernateTopicTestSupport {
             deleteDummyEntity(sf, i);
         }
 
-        assertTopicNotifications(6, CACHE_ENTITY);
+        assertTopicNotifications(3, CACHE_ENTITY);
         assertTopicNotifications(3, CACHE_ENTITY_PROPERTIES);
-        assertTopicNotifications(24, CACHE_PROPERTY);
+        assertTopicNotifications(12, CACHE_PROPERTY);
         assertTopicNotifications(67, CACHE_TIMESTAMPS_REGION);
     }
 
-    @Test(expected = UnsupportedOperationException.class)
+    @Test
     public void testUpdateOneEntity() {
         insertDummyEntities(sf, 10, 4);
 
         getDummyEntities(sf, 10);
 
         updateDummyEntityName(sf, 2, "updated");
+
+        assertTopicNotifications(1, CACHE_ENTITY);
+        assertTopicNotifications(0, CACHE_ENTITY_PROPERTIES);
+        assertTopicNotifications(0, CACHE_PROPERTY);
+        assertTopicNotifications(54, CACHE_TIMESTAMPS_REGION);
     }
 
-    // This test should throw an UnsupportedOperationException, for attempting an update on a read-only cache,
-    // but in Hibernate 5.3 the exception is not thrown.
-    @Ignore
-    @Test(expected = UnsupportedOperationException.class)
+    @Test
     public void testUpdateEntities() {
         insertDummyEntities(sf, 1, 10);
 
         executeUpdateQuery(sf, "update DummyEntity set name = 'updated-name' where id < 2");
-    }
 
-    // This test should throw an UnsupportedOperationException, for attempting an update on a read-only cache,
-    // but in Hibernate 5.3 the exception is not thrown.
-    @Ignore
-    @Test(expected = UnsupportedOperationException.class)
-    public void testUpdateEntitiesAndProperties() {
-        insertDummyEntities(sf, 1, 10);
-
-        Session session = null;
-        Transaction txn = null;
-        try {
-            session = sf.openSession();
-            txn = session.beginTransaction();
-            Query query = session.createQuery( "update DummyEntity set name = 'updated-name' where id < 2");
-            query.setCacheable(true);
-            query.executeUpdate();
-
-            Query query2 = session.createQuery( "update DummyProperty set version = version + 1");
-            query2.setCacheable(true);
-            query2.executeUpdate();
-
-            txn.commit();
-        } catch (RuntimeException e) {
-            txn.rollback();
-            e.printStackTrace();
-            throw e;
-        } finally {
-            session.close();
-        }
-    }
-
-    // This test should throw an UnsupportedOperationException, for attempting an update on a read-only cache,
-    // but in Hibernate 5.3 the exception is not thrown.
-    @Ignore
-    @Test(expected = UnsupportedOperationException.class)
-    public void testUpdateOneEntityAndProperties() {
-        insertDummyEntities(sf, 1, 10);
-
-        Session session = null;
-        Transaction txn = null;
-        try {
-            session = sf.openSession();
-            txn = session.beginTransaction();
-            Query query = session.createQuery( "update DummyEntity set name = 'updated-name' where id = 0");
-            query.setCacheable(true);
-            query.executeUpdate();
-
-            Query query2 = session.createQuery( "update DummyProperty set version = version + 1");
-            query2.setCacheable(true);
-            query2.executeUpdate();
-
-            txn.commit();
-        } catch (RuntimeException e) {
-            txn.rollback();
-            e.printStackTrace();
-            throw e;
-        } finally {
-            session.close();
-        }
-    }
-
-    protected String getCacheStrategy() {
-        return defaultAccessType.getExternalName();
+        assertTopicNotifications(1, CACHE_ENTITY);
+        assertTopicNotifications(0, CACHE_ENTITY_PROPERTIES);
+        assertTopicNotifications(0, CACHE_PROPERTY);
+        assertTopicNotifications(15, CACHE_TIMESTAMPS_REGION);
     }
 
     @Override
-    protected Properties getCacheProperties() {
-        Properties props = new Properties();
-        props.put("TestAccessType", defaultAccessType);
-        props.setProperty(Environment.CACHE_REGION_FACTORY, HazelcastLocalCacheRegionFactory.class.getName());
-        return props;
+    protected AccessType getCacheStrategy() {
+        return AccessType.READ_WRITE;
     }
 }
